[REFACTOR] Post 변경 및 생성 시, userId 로직 수정

+++ src/main/java/com/example/backend_webflux/service/PostService.java
@@ -34,11 +34,11 @@ public class PostService {
         .switchIfEmpty(Mono.error(new ApiException(ApiExceptionEnum.NOT_FOUND_EXCEPTION)));
   }
 
-  public Mono<Post> modify(String id, PostDto dto) {
+  public Mono<Post> modify(String id, String userId, PostDto dto) {
     return postRepository.findById(id)
         .switchIfEmpty(Mono.error(new ApiException(ApiExceptionEnum.NOT_FOUND_EXCEPTION)))
         .flatMap(result -> {
-          if (result.getUser().equals(dto.getUserId())) {
+          if (result.getUser().equals(userId)) {
             result.setTitle(dto.getTitle());
             result.setContent(dto.getContent());
             result.setStatus(1);
@@ -49,12 +49,18 @@ public class PostService {
         });
   }
 
-  public Mono<Void> delete(String id) {
-    return postRepository.deleteById(id);
+  public Mono<Void> delete(String id, String userId) {
+    return userRepository.findById(userId)
+        .flatMap(writer -> postRepository.findById(id).flatMap(post -> {
+          if( writer.getId().equals(post.getUser())) {
+            return postRepository.deleteById(id);
+          }
+          return Mono.empty();
+        })).switchIfEmpty(Mono.error(new ApiException(ApiExceptionEnum.WRITER_UNAUTHORIZED_EXCEPTION)));
   }
 
-  public Mono<Post> create(PostDto dto) {
-    return userRepository.findById(dto.getUserId())
+  public Mono<Post> create(String userId, PostDto dto) {
+    return userRepository.findById(userId)
         .switchIfEmpty(Mono.error(new ApiException(ApiExceptionEnum.NOT_FOUND_EXCEPTION)))
         .flatMap(user -> {
           Post p = new Post();
